Move file provider types to new folder
Prep for #4386
